initiate

+++ market/src/main/java/com/skydive/market/model/registration/RegistrationDTO.java
@@ -13,7 +13,7 @@ public class RegistrationDTO {
     private String lastName;
     private String dob;
     private String email;
-    private Integer phoneNumber;
+    private Long phoneNumber;
     private String country;
     private boolean isSportsMan;
     private SportType typeOfSport;
